example: migrated to providers file over wagmi and rainbowkit config in layout
